Prior patch for issue2550826 captured too many exceptions from detectors. Including those exceptions that have to be passed up unmolested: ValueError, Reject, RejectRaw.
 This broke a number of tests in

  test/test_userauditor.py

This patch passes these excpetions up the chain bypassing the
DetectorError exception. It alo include pytz's UnknownTimezoneError
which is passed up by an auditor.

This isn't the greatest patch, but it does fix the test suite.  There
may be another set of patches in this area depending on the answers I
get from my mail to the roundup-devel mailinglist.