Update run_tests.py to pytest v2.9.2
pytest v2.9.2 fixes an issue where 'pytest.mark.skip' would not
correctly skip tests.

There was some discussion on the mailing list about removing the
run_test.py script in favour of having pytest as an external dependency,
but for the time being updating the script is the quicker/easier option.