issue2550727: db.newid is broken with sqlite.
Added proper transaction lock around the sql code to get a new id. The
the locking that pysqlite attempts had to be defeated because it is
broken.  Had to explicitly manage transactions with BEGIN IMMEDIATE
and call sql_commit.

Note that this reduces performance by 30% in return for accuracy.

Also use update call set newid=newid+1 rather than incrementing in
python.