Add support for prev/next/self links when returning paginated results.
To do this:
  1) change "data" envelope from an array to a dict
  2) move the "data" array to the "collection" property,
      which is an array of elements in the collection.
  3) add @links dict keyed by link relation: self, next, prev.
     Each relation is an array of dicts with uri and rel keys.
     In this case there is only one element, but there is nothing
     preventing a relation from having multiple url's. So this follows
     the formatting needed for the general case.
     Relations are present only if it makes sense. So first page has no
     prev and last page has no next.
  4) add @total_size with number of element selected if they were
     not paginated. Replicates data in X-Count-Total header.

Changed index to start at 1. So the first page is page_index 1 and not
page_index 0. (So I am no longer surprised when I set page_index to 1
and am missing a bunch of records 8-)).

Also a small fixup, json response ends with a newline so printing
the data, or using curl makes sure that anything printing after the
json output (like shell prompts) is on a new line.

Tests added for all cases.