Fix REST API for WSGI
Add dictionary-like headers attribute.
Similar to the apache implementation in same directory but we look up
the given header in the wsgi environment. Environment entries are
mangled (all uppercase, '-' replaced with '_', prefix HTTP_). So on
lookup we mangle the given name first. This allows the rest
implementation (and others) to look up headers.